requirement.txt
Flask==2.3.3
flask-cors==4.0.0
openai==0.27.8
requests==2.31.0
Pillow==10.0.0
python-dotenv==1.0.0